# Implement new evaluation of multi-pathway dipolar models to greatly improve performance