I have been profiling DeerLab's code after merging #390, particularly on analyses based on dipolar EPR models created with `dipolarmodel`. The bottleneck of the execution of the `fit` function for any dipolar model is the execution of the elementary dipolar kernel matrix in `dipolarkernel` with most of the runtime spent on the `elementarykernel_fresnel` (or another function, such as `elementarykernel_grid` depending on the construction method of the kernel) or `memoization.caching._lru_cache` both inside of `elementarykernel_twospin`. 

The reason for such a bottleneck lies in the nature of multi-pathway dipolar models. Each dipolar pathway has a different effective dipolar evolution time (due to having a different refocusing time). Therefore, for each pathway, the `scipy.special.fresnel` function has to be evaluated. Despite being a generally fast function, the hundreds of execution required during a typical execution of the `fit` function turn out to be a bottleneck. While memorization helps, due to the constant changing of refocusing times values during a fit, the `dipolarkernel` function has to recompute a new pathway not cached, which is then stored in the cache. What ends up happening is that the cache is being constantly checked and filled, reaching maximum capacity after 100 pathway evaluations, and then starting to maintain the cache, which ends up costing time. 

## Features 

This PR aims to unclog (or at least improve) this bottleneck. It implements a new system to evaluate dipolar pathways. Since all dipolar pathways share the same dipolar kernel, with the only difference is being time-shifted with respect to each other, we can envision the construction of a very large dipolar kernel from which we only need to take cutouts for the different pathways depending on their refocusing times. 

The new implementation achieves this by means of interpolation. It adds a new optional argument `tinterp` to `dipolarkernel`. This argument takes such a large time vector which is used to build a dipolar kernel larger than the one that would be built from `t`. This required now a first execution of  `elementarykernel_twospin`. Now, if `tinterp` covers all possible refocusing-time-shifted vectors`t - tref` then we can construct all their dipolar kernels by interpolation from that first large kernel (very accurately) without the need to re-run the `elementarykernel_twospin` function.

Now, in `dipolarmodel` when specifying the `ex_` models we know exactly the refocusing times and their boundaries. Thus, we can know exactly the minimal `tinterp` vector covering all possible refocusing times possibilities. By exploiting that, we can evaluate the dipolar model such that one instance of a dipolar model only needs to execute `elementarykernel_twospin` once independently of the number of pathways. 

When executing the `fit` function, the dipolar model instance will be evaluated hundreds or thousands of times.  However, thanks to the memorization of `dipolarkernel` and `tinterp` being kept constant over the `fit` routine, the costly `elementarykernel_twospin` function will not be executed anymore. Therefore, with this new system, the dipolar kernel is only constructed explicitly once over the full analysis, significantly reducing the computation time.

Current limitations:
- For now limited to models where the `experiment` argument is specified in `dipolarmodel`.
- Sensible to the user changing the boundaries of the refocusing times after construction, such that `tinterp` might not cover all possibilities making interpolation unfeasible. To account for this possibility, the program will now detect such cases and roll back to the previous mode. 


## Benchmark

I benchmarked the performance of the `fit` function when analyzing 4-pulse DEER data (`example_4pdeer_2.DTA`, 488 data points) and 5-pulse DEER data (`example_5pdeer_3.DTA`, 926 data points) with a different number of dipolar pathways. Due to its robustness, I also included a benchmark when enforcing the compactness penalty upon the underlying distance distribution. The benchmark results are shown for all these cases with (`Now`) and without (`Before`) speedup. In all cases I double-checked that I would get the same results with and without the speedup to ensure that the accuracy remains unchanged.

You can find the benchmark attached [here](https://github.com/JeschkeLab/DeerLab/files/9882072/benchmark_multipathway_deer_speedup.zip). The results are the following:

#### 4-pulse DEER models
| Pathways | Before  (without compactness penalty |   Now  (with compactness penalty)   | Before (with compactness penalty) |    Now (with compactness penalty)    |
|:--------:|:------------------------------------:|:-----------------------------------:|:---------------------------------:|:------------------------------------:|
|     1    |               **0.79s**              |           1.09s (:x: +40%)          |             **11.25s**            |           14.01s (:x: +24%)          |
|     2    |                 1.79s                | **1.34s** (:white_check_mark: -25%) |               19.88s              | **15.84s** (:white_check_mark: -20%) |
|     3    |                 3.08s                | **1.79s** (:white_check_mark: -42%) |               35.06s              | **20.39s** (:white_check_mark: -42%) |
|     4    |                 4.42s                | **2.09s** (:white_check_mark: -53%) |               39.09s              | **32.83s** (:white_check_mark: -16%) |

#### 5-pulse DEER models
| Pathways | Before  (without compactness penalty |   Now  (with compactness penalty)  | Before (with compactness penalty) |    Now (with compactness penalty)   |
|:--------:|:------------------------------------:|:----------------------------------:|:---------------------------------:|:-----------------------------------:|
|     1    |                 1.85s                |  **1.10s** (:white_check_mark: -41%) |               18.92s              |  **13.47s** (:white_check_mark: -29%) |
|     2    |                 3.04s                |  **1.75s** (:white_check_mark: -42%) |               36.61s              |  **23.89s** (:white_check_mark: -35%) |
|     3    |                 5.58s                |  **2.38s** (:white_check_mark: -57%) |               80.20s              |  **29.77s** (:white_check_mark: -63%) |
|     4    |                 7.16s                |  **2.39s** (:white_check_mark: -66%) |              100.58s              |  **33.66s** (:white_check_mark: -66%) |
|     5    |                13.59s                |  **5.08s** (:white_check_mark: -63%) |              132.34s              |  **44.95s** (:white_check_mark: -66%) |
|     6    |                15.92s                |  **5.52s** (:white_check_mark: -65%) |              198.26s              |  **81.18s** (:white_check_mark: -59%) |
|     7    |                29.34s                | **13.41s** (:white_check_mark: -54%) |              318.90s              |  **83.25s** (:white_check_mark: -74%) |
|     8    |                32.08s                |  **9.44s**( :white_check_mark: -70%) |              362.20s              | **126.25s** (:white_check_mark: -65%) | 

The results show how this implementation can provide massive speedups for multi-pathway analyses. I have only observed a detrimental behavior for 4-pulse DEER (i.e., relatively low data points) with a single pathway. While the loss is not substantial (due to the overall short execution of such simple models) one could also hardcode the speedup to be disabled for single-pathway models by default.

When combined with the performance gained with PR #390 we are looking at an **average reduction of execution time for dipolar EPR analyses with DeerLab of 70-80%** :rocket:.