@luisfabib 2022-10-27
Another illustrative benchmark: All tests pass in **5.90min** rather than 10.18min (from the last merged PR).